requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
Django==4.1.3
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
flake8==5.0.4
identify==2.5.9
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==22.0
pathspec==0.10.2
platformdirs==2.5.4
pre-commit==2.20.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.28.1
sentry-sdk==1.12.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
vine==5.0.0
virtualenv==20.16.7
wcwidth==0.2.5
